refactor: (#515) score 반환 long으로 변환

+++ backend/src/main/java/com/votogether/domain/member/dto/response/RankingResponse.java
@@ -17,7 +17,7 @@ public record RankingResponse(
         int voteCount,
 
         @Schema(description = "점수", example = "31")
-        int score
+        long score
 ) {
 }
